test: 카드 합 구하는 테스트 위치 이동
